# Conflicts:
#	lib/hsa/mcwamp_hsa.cpp

+++ include/hc_defines.h
@@ -30,19 +30,19 @@ namespace hc {
 //
 // work-item related builtin functions
 //
-extern "C" __attribute__((const,hc)) int64_t hc_get_grid_size(unsigned int n);
-extern "C" __attribute__((const,hc)) int64_t hc_get_workitem_absolute_id(unsigned int n);
-extern "C" __attribute__((const,hc)) int64_t hc_get_group_size(unsigned int n);
-extern "C" __attribute__((const,hc)) int64_t hc_get_workitem_id(unsigned int n);
-extern "C" __attribute__((const,hc)) int64_t hc_get_num_groups(unsigned int n);
-extern "C" __attribute__((const,hc)) int64_t hc_get_group_id(unsigned int n);
+extern "C" __attribute__((const,hc)) uint32_t hc_get_grid_size(unsigned int n);
+extern "C" __attribute__((const,hc)) uint32_t hc_get_workitem_absolute_id(unsigned int n);
+extern "C" __attribute__((const,hc)) uint32_t hc_get_group_size(unsigned int n);
+extern "C" __attribute__((const,hc)) uint32_t hc_get_workitem_id(unsigned int n);
+extern "C" __attribute__((const,hc)) uint32_t hc_get_num_groups(unsigned int n);
+extern "C" __attribute__((const,hc)) uint32_t hc_get_group_id(unsigned int n);
 
-extern "C" __attribute__((const,amp)) int64_t amp_get_global_size(unsigned int n);
-extern "C" __attribute__((const,amp)) int64_t amp_get_global_id(unsigned int n); 
-extern "C" __attribute__((const,amp)) int64_t amp_get_local_size(unsigned int n);
-extern "C" __attribute__((const,amp)) int64_t amp_get_local_id(unsigned int n);
-extern "C" __attribute__((const,amp)) int64_t amp_get_num_groups(unsigned int n);
-extern "C" __attribute__((const,amp)) int64_t amp_get_group_id(unsigned int n);
+extern "C" __attribute__((const,amp)) uint32_t amp_get_global_size(unsigned int n);
+extern "C" __attribute__((const,amp)) uint32_t amp_get_global_id(unsigned int n);
+extern "C" __attribute__((const,amp)) uint32_t amp_get_local_size(unsigned int n);
+extern "C" __attribute__((const,amp)) uint32_t amp_get_local_id(unsigned int n);
+extern "C" __attribute__((const,amp)) uint32_t amp_get_num_groups(unsigned int n);
+extern "C" __attribute__((const,amp)) uint32_t amp_get_group_id(unsigned int n);
 
 #if __KALMAR_ACCELERATOR__ == 2
 #define tile_static thread_local
